Add Tidepool navigation view - One more click to reach the service, but - Added chevron in settings list (for consistency) - Added room for information/explainations of the setting in header/footer in the new view - First draft to build on (maybe: rephrase header/footer, add connection status in navigation view, other relevant info etc)
